MAPR-HIVE-556 : Error while starting Hive CLI on Tez: ERROR exec.Task: Failed to execute tez graph

+++ ql/src/java/org/apache/hadoop/hive/ql/exec/tez/TezSessionState.java
@@ -435,9 +435,11 @@ public class TezSessionState {
    * @throws IOException
    */
   private void saveActiveAppToFile() throws IOException {
-    FileSystem fs = activeAppsPath.getFileSystem(conf);
-    FSDataOutputStream fsDataOutputStream = fs.create(new Path(activeAppsPath, applicationId.toString()));
-    fsDataOutputStream.close();
+    if (activeAppsPath != null) {
+      FileSystem fs = activeAppsPath.getFileSystem(conf);
+      FSDataOutputStream fsDataOutputStream = fs.create(new Path(activeAppsPath, applicationId.toString()));
+      fsDataOutputStream.close();
+    }
   }
 
   /**
@@ -446,8 +448,8 @@ public class TezSessionState {
    * @throws IOException
    */
   private void deleteFileWithAppId() throws IOException {
-    FileSystem fs = activeAppsPath.getFileSystem(conf);
-    if (applicationId != null) {
+    if (activeAppsPath != null && applicationId != null) {
+      FileSystem fs = activeAppsPath.getFileSystem(conf);
       Path jobIdPath = new Path(activeAppsPath, applicationId.toString());
       if (fs.exists(jobIdPath)) {
         fs.delete(jobIdPath, true);
